ai/STT/requirements.txt
openai-whisper==20231117
numpy<2.3
